requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.23.0
supabase==1.0.3
apscheduler==3.10.1
python-dotenv==1.0.0
